Adjust times for BST.
This changes a lot of values that seem like they shouldn't change. The
data runs from July 2013 in clock time, which means that the first few
entries were in June GMT. When we merged the bikes with the weather,
those entries were dropped because we have no weather data for them. Now
we're not ignoring those entries, which means a lot of things change
slightly.

(We now explicitly drop the leading entries, with NAs in prev.* and d.*,
because they were showing up in some outputs as all-NA rows. (That
happens if you do df[filter,] and filter contains an NA.))

This also changes some plots and values significantly. I haven't looked
hard for them yet, but I've noticed the R^2 values have reduced.
Apparently there's less correlation 0900-0930 than 1000-1030.